requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.9
Django==5.0.14
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
environ==1.0
filelock==3.17.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.5.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.51
psycopg==3.2.5
psycopg2==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.29.2
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
